Implement reduction of ifcmpge to ifcmpeq
In specific cases, we can reduce the ifcmpge opcode to ifcmpeq.
Once this transformation is performed, the test-under-mask optimization
can be applied to the "ifcmpeq" node.

The reduction is possible on the following tree:
ificmpge --> block_5
 iand
   iload  <...>
   iconst 16
 iconst 16
Since the constants are equal, the "greater-than" case is impossible and
ifcmpeq can be used instead.

Issue: #1556

Signed-off-by: Alexey Zhikhartsev <alexey.zhikhartsev@ibm.com>